src/e/gui/ArrayListTableModel.java: remove unused (and not particularly useful) class.
src/e/gui/DebugMenu.java: 
src/e/gui/HeapView.java: 
src/e/gui/HorizontalScrollWheelListener.java: 
src/e/ptextarea/PTextAreaRenderer.java: 
src/e/tools/JarExplorer.java: 
src/e/tools/JavaHpp.java: 
src/e/tools/ShowUiDefaults.java: 
src/e/util/InAppServer.java: 
src/e/util/Launcher.java: 
src/e/util/Log.java: 
src/e/util/Stopwatch.java: fix raw type, unused local/field/method, and unnecessary cast warnings.

+++ salma-hayek/trunk/src/e/util/Stopwatch.java
@@ -62,7 +62,7 @@ public class Stopwatch {
         if (sampleCount == 0) {
             result += " (no samples)";
         } else {
-            result += StringUtilities.pluralize(sampleCount, "sample", "samples") + ", " + nsToString(totalDuration_ns) + " total, " + nsToString(minDuration_ns) + ".." + nsToString(maxDuration_ns) + " (mean " + nsToString(totalDuration_ns/(long) sampleCount) + ")";
+            result += StringUtilities.pluralize(sampleCount, "sample", "samples") + ", " + nsToString(totalDuration_ns) + " total, " + nsToString(minDuration_ns) + ".." + nsToString(maxDuration_ns) + " (mean " + nsToString(totalDuration_ns/sampleCount) + ")";
         }
         return result;
     }
